Precompute cell values for the regular cell neighboring type
1. Implemented the said mechanism for the three code branches: full matrix assembly, H-matrix assembly without the producer-consumer model and task buffers, H-matrix assembly with the producer-consumer model and task buffers.
2. Refactored the code structure by making some low level Sauter quadrature related functions platform shared, namely,
   - ~transform_quad_point_from_unit_to_permuted_real_cell~
   - ~surface_jacobian_det_and_normal_vector~
   - ~surface_jacobian_det_normal_vector_and_covariant~
   - ~surface_curl~
3. Refactored the code for BEM values and scratch data:
   - Instead of using a common scratch data class ~PairCellWiseScratchData~ for all code branches, now use different classes.
   - Removed redundant data table allocation in BEM values and scratch data.
4. Preliminary verification of test cases.

+++ include/build_tools/template_instantiator.h
@@ -48,7 +48,8 @@
     TI_INSTANTIATE_CLS_4_WITH_RANGE_KERNEL_NUMBER_TYPES_,                        \
     (cls_seq)(arg1_seq)(arg2_seq)(arg3_seq))
 
-#define TI_INST_NUM_TYPES (float)(double)
+// #define TI_INST_NUM_TYPES (float)(double)
+#define TI_INST_NUM_TYPES (double)
 #define TI_INST_SPACE_DIMS (3)
 #define TI_INST_BOUNDARY_DIMS (2)
